slmember

+++ src/main/java/com/sy/yuanfei/mapper/AuFunctionMapper.java
@@ -4,5 +4,5 @@ import com.sy.yuanfei.model.AuFunction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface AuFunctionMapper extends JpaRepository<AuFunction,Long> {
+public interface AuFunctionMapper extends JpaRepository<AuFunction,Integer> {
 }
